fix(arm32): fix ARM32 cross-build compatibility and unaligned access crashes
Add ARM32 CMake toolchain file that auto-detects qemu-arm-static and
passes -L flag directly, replacing fragile /tmp wrapper script approach.

Fix SIGBUS (unaligned uint64_t access) in packed structs on ARM32 by
replacing direct pointer casts with memcpy in:
- dap_chain_block.c: block meta NONCE/NONCE2 and EXCLUDED_KEYS parsing
- dap_chain_datum_tx_voting.c: voting TSD EXPIRE/MAX_VOTES_COUNT parsing
- dap_chain_cs_esbocs.c: SYNC_ATTEMPT block metadata extraction
- dap_chain_ch.c: chain ACK packet number reading
- dap_chain_node_dns_client.c: DNS response address parsing
- dap_chain_ledger.c: UTXO blocking TSD timestamp parsing
- dap_chain_net_srv_voting.c: voting params TSD parsing

Fix format specifier mismatches (ARM32 type size differences) in:
- dap_chain_cell.c, dap_chain_cs_dag.c: %lu -> %zu for size_t
- dap_stream_ch_chain_net_srv.c: %ld -> %lld for intmax_t
- dap_chain_net_srv_vpn.c: %ld/%lu -> %lld/%zd/%zu for intmax_t/size_t
- dap_chain_net_srv_vpn_cmd.c: %ld -> %llu/%lld for uintmax_t/intmax_t
- dap_chain_node_cli_cmd.c: %lu -> %lld for time_t in snprintf
- dap_chain_ledger.c: %lu -> %zu for size_t

Fix use-after-free in arbitrage_cli_integration_test.c: copy JSON hash
string to stack buffer before freeing the JSON object.

Bump dap-sdk submodule to include ARM32-safe XKCP Keccak implementation.

All 12 ARM32 tests pass (100%).

+++ modules/net/dap_chain_node_cli_cmd.c
@@ -2174,7 +2174,7 @@ int l_arg_index = 1, l_rc, cmd_num = CMD_NONE;
                     }
                     // create wallet backup 
                     dap_chain_wallet_internal_t* l_file_name = DAP_CHAIN_WALLET_INTERNAL(l_wallet);
-                    snprintf(l_file_name->file_name, sizeof(l_file_name->file_name), "%s/%s_%012lu%s", c_wallets_path, l_wallet_name, time(NULL),".backup");
+                    snprintf(l_file_name->file_name, sizeof(l_file_name->file_name), "%s/%s_%012lld%s", c_wallets_path, l_wallet_name, (long long)time(NULL),".backup");
                     if ( dap_chain_wallet_save(l_wallet, NULL) ) {
                         dap_json_rpc_error_add(*a_json_arr_reply, DAP_CHAIN_NODE_CLI_COM_TX_WALLET_BACKUP_ERR,
                                                "Can't create backup wallet file because of internal error");
